week_3/requirements.txt
streamlit
python-dotenv
openai
groq
chromadb
sentence-transformers
langchain==0.2.16
langchain-community==0.2.12
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
rank-bm25
youtube-transcript-api
pytube
trafilatura
readability-lxml
beautifulsoup4
requests
pydub
gTTS
yt-dlp
beautifulsoup4
lxml
